wip: fix MDX slugs by only creating slug and timeToRead fields for MDX nodes in the /blog/ directory
Still getting the following error (and a page created for route /blog/{mdx.fields__slug})

 ERROR #gatsby-plugin-page-creator_12104  PLUGIN

PageCreator: Could not find a value in the node for blog/{mdx.fields__slug}.js. Please make
sure that the syntax is correct and supported.

See our docs page for more info on this error:
https://www.gatsbyjs.com/docs/file-system-route-api/\#syntax-collection-routes